backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
ibm-watson==8.1.0
ibm-cloud-sdk-core==3.20.1